backend/services/account/requirements.txt
uvicorn==0.38.0
aio-pika==9.5.8
fastapi[standard]==0.121.2
grpcio==1.76.0
grpcio-tools==1.76.0
python-dotenv==1.2.1
sqlalchemy==2.0.44
PyJWT==2.10.1
jdatetime==5.2.0
phonenumbers==9.0.18
bcrypt==4.2.0
asyncpg==0.30.0
httpx==0.27.2